+++ CHANGES.rst
@@ -2,9 +2,15 @@ Unreleased
 ----------
 - Enable CI and add tests [#226]
 
+- When computing an orthogonal CD matrix, check that ``idcscale`` exists
+  and is not ``None``. If ``idcscale`` is missing (some headerlets in
+  the Astrometry DB), compute a scale from the CD matrix. [#227]
+
+- Return default response when request to the GSC service failed. [#227]
+
 1.7.5 (2025-07-14)
 ------------------
-- When the GSC web service returns no RA and DEC offsets from GAIA (0.), use default FITS WCS values.[#221]
+- Work around an issue with the AStrometry DB which returns NaNs in certain cases. [#221]
 
 1.7.4 (2025-01-07)
 ------------------